requirements.txt
autopep8==2.0.2
jsbeautifier==1.14.9
openai==0.27.8
Pillow==10.0.0
PyAudio==0.2.13
python-decouple==3.8
pyttsx3==2.90
pywhatkit==5.4
shortuuid==1.0.11
SpeechRecognition==3.10.0
